requirements.txt
fastapi==0.115.12
joblib==1.4.2
nltk==3.9.1
numpy==2.2.5
opencv_python==4.11.0.86
Pillow==11.2.1
pydantic==2.11.4
tensorflow==2.16.1
tensorflow_intel==2.16.1
